+++ .circleci/continue_config.yml
@@ -25,7 +25,7 @@ parameters:
   run-build-cci-stats:
     type: boolean
     default: false
-  run-build-op-nat:
+  run-build-op-acceptor:
     type: boolean
     default: false
   run-all:
@@ -96,7 +96,7 @@ jobs:
             echo "run-build-pms: << pipeline.parameters.run-build-pms >>"
             echo "run-build-op-ufm: << pipeline.parameters.run-build-op-ufm >>"
             echo "run-build-proxyd: << pipeline.parameters.run-build-proxyd >>"
-            echo "run-build-op-nat: << pipeline.parameters.run-build-op-nat >>"
+            echo "run-build-op-acceptor: << pipeline.parameters.run-build-op-acceptor >>"
             echo "run-all: << pipeline.parameters.run-all >>"
             echo ""
             echo "Pipeline Trigger Information:"
@@ -143,10 +143,10 @@ jobs:
               echo "op-ufm tag regex match: false"
             fi
 
-            if [[ $CURRENT_TAG =~ ^op-nat/v.* ]]; then
-              echo "op-nat tag regex match: true"
+            if [[ $CURRENT_TAG =~ ^op-acceptor/v.* ]]; then
+              echo "op-acceptor tag regex match: true"
             else
-              echo "op-nat tag regex match: false"
+              echo "op-acceptor tag regex match: false"
             fi
   
   docker-build:
@@ -519,20 +519,20 @@ workflows:
           docker_name: cci-stats
           docker_tags: <<pipeline.git.revision>>,<<pipeline.git.branch>>
           docker_context: .
-  op-nat:
+  op-acceptor:
     when:
-      or: [<< pipeline.parameters.run-build-op-nat >>, << pipeline.parameters.run-all >>]
+      or: [<< pipeline.parameters.run-build-op-acceptor >>, << pipeline.parameters.run-all >>]
     jobs:
       - go-lint:
-          name: op-nat-lint
-          module: op-nat
+          name: op-acceptor-lint
+          module: op-acceptor
       - go-test:
-          name: op-nat-tests
-          module: op-nat
+          name: op-acceptor-tests
+          module: op-acceptor
       - docker-build:
-          name: op-nat-docker-build
-          docker_file: op-nat/Dockerfile
-          docker_name: op-nat
+          name: op-acceptor-docker-build
+          docker_file: op-acceptor/Dockerfile
+          docker_name: op-acceptor
           docker_tags: <<pipeline.git.revision>>,<<pipeline.git.branch>>
           docker_context: .
   release:
@@ -540,14 +540,14 @@ workflows:
       - log-config-results:
           filters:
             tags:
-              only: /^(op-nat|cci-stats|peer-mgmt-service|proxyd|ufm-[a-z0-9\-]*|op-[a-z0-9\-]*)\/v.*/
+              only: /^(cci-stats|peer-mgmt-service|proxyd|ufm-[a-z0-9\-]*|op-[a-z0-9\-]*)\/v.*/
             branches:
               ignore: /.*/
       - hold:
           type: approval
           filters:
             tags:
-              only: /^(op-nat|cci-stats|peer-mgmt-service|proxyd|ufm-[a-z0-9\-]*|op-[a-z0-9\-]*)\/v.*/
+              only: /^(cci-stats|peer-mgmt-service|proxyd|ufm-[a-z0-9\-]*|op-[a-z0-9\-]*)\/v.*/
             branches:
               ignore: /.*/
       - docker-build:
@@ -561,7 +561,7 @@ workflows:
                 - op-conductor-mon
                 - peer-mgmt-service
                 - cci-stats
-                - op-nat
+                - op-acceptor
           name: <<matrix.docker_name>>-docker-build
           filters:
             tags:
@@ -584,7 +584,7 @@ workflows:
                 - op-conductor-mon
                 - peer-mgmt-service
                 - cci-stats
-                - op-nat
+                - op-acceptor
           name: <<matrix.docker_name>>-docker-publish
           filters:
             tags:
@@ -605,7 +605,7 @@ workflows:
                 - op-conductor-mon
                 - peer-mgmt-service
                 - cci-stats
-                - op-nat
+                - op-acceptor
           name: <<matrix.docker_name>>-docker-tag
           filters:
             tags:
@@ -627,7 +627,7 @@ workflows:
                 - op-conductor-mon
                 - peer-mgmt-service
                 - cci-stats
-                - op-nat
+                - op-acceptor
           name: <<matrix.module>>-go-release
           filters:
             tags:
