+++ op-acceptor/Makefile
@@ -10,10 +10,10 @@ LDFLAGS := -ldflags "$(LDFLAGSSTRING)"
 all: build
 
 build:
-	go build -v $(LDFLAGS) -o ./bin/op-nat ./cmd
+	go build -v $(LDFLAGS) -o ./bin/op-acceptor ./cmd
 
 clean:
-	rm ./bin/op-nat
+	rm ./bin/op-acceptor
 
 test:
 	go test -v ./...
