+++ op-acceptor/manifest.go
