+++ op-acceptor/nat_test.go
