+++ op-acceptor/runner/runner.go
@@ -10,9 +10,9 @@ import (
 
 	"errors"
 
-	"github.com/ethereum-optimism/infra/op-nat/metrics"
-	"github.com/ethereum-optimism/infra/op-nat/registry"
-	"github.com/ethereum-optimism/infra/op-nat/types"
+	"github.com/ethereum-optimism/infra/op-acceptor/metrics"
+	"github.com/ethereum-optimism/infra/op-acceptor/registry"
+	"github.com/ethereum-optimism/infra/op-acceptor/types"
 	"github.com/ethereum/go-ethereum/log"
 	"github.com/google/uuid"
 )
