+++ op-acceptor/types/gate.go
